+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - Fixed Refined Storage jars not being signed (raoulvdberge)
 - Fixed client stalling when trying to search with # for tooltips (raoulvdberge)
 - Fixed crafting task stalling when there's not enough space in the inventory (raoulvdberge)
+- Fixed another duplication bug with a disconnected Crafting Grid (raoulvdberge)
 - Removed getMissingItems() and getMissingFluids() functions from the OpenComputers integration, that info is now accessible through schedule(Fluid)Task(). If you just want to check if there are missing items/fluids but don't want to start an actual task, use the "canSchedule" parameter (raoulvdberge)
 - Updated Russian translation (kellixon)
 - Added fluid functions for the fluid autocrafting to the OpenComputers integration (raoulvdberge)